+++ src/components/forms/edit-account.tsx
@@ -6,10 +6,10 @@ import filter from 'just-filter-object'
 import { TFunction, useTranslation } from 'next-i18next'
 import * as yup from 'yup'
 
+import { UpdateUserDto, User } from '@/modules/auth/auth'
 import { useClubsList } from '@/modules/clubs/hooks'
 import { useRegionsList } from '@/modules/regions/hooks'
 import { useUserFootballRolesList } from '@/modules/user-football-roles/hooks'
-import { UpdateUserDto, User } from '@/types/auth'
 import { validateId } from '@/utils/validation-helpers'
 
 import { ClubsCombo } from '../selects/clubs-combo'
