+++ src/components/forms/utils.ts
@@ -1,7 +1,7 @@
 import { TFunction } from 'next-i18next'
 import * as yup from 'yup'
 
-import { LoginDto } from '@/types/auth'
+import { LoginDto } from '@/modules/auth/auth'
 
 export function generatePasswordValidationSchema(t: TFunction) {
   return yup
