+++ src/components/nav/logout-button.tsx
@@ -2,7 +2,7 @@ import { ExitToApp as LogoutIcon } from '@mui/icons-material'
 import { ListItemButton, ListItemIcon, ListItemText } from '@mui/material'
 import { useTranslation } from 'next-i18next'
 
-import { useLogout } from '@/lib/auth'
+import { useLogout } from '@/modules/auth/hooks'
 
 export const LogoutButton = () => {
   const { t } = useTranslation()
