+++ src/pages/api/logout.ts
@@ -1,6 +1,6 @@
 import { withIronSessionApiRoute } from 'iron-session/next'
 
-import { sessionOptions } from '@/lib/session'
+import { sessionOptions } from '@/modules/auth/session'
 
 export default withIronSessionApiRoute((req, res) => {
   req.session.destroy()
