+++ src/pages/clubs/edit/[slug].tsx
@@ -5,7 +5,7 @@ import { ErrorContent } from '@/components/error/error-content'
 import { EditClubForm } from '@/components/forms/club/edit-club'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useUpdateClub } from '@/modules/clubs/hooks'
 import { ClubDto } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
