+++ src/pages/clubs/index.tsx
@@ -10,7 +10,7 @@ import { ConfirmationModal } from '@/components/modals/confirmation-modal'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { ClubsTable } from '@/components/tables/clubs'
 import { ClubsTableRow } from '@/components/tables/rows/clubs-row'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useClubs, useDeleteClub } from '@/modules/clubs/hooks'
 import { ClubsFiltersDto, ClubsSortBy } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
