+++ src/pages/dashboard.tsx
@@ -1,6 +1,6 @@
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import { withSessionSsr } from '../lib/session'
+import { withSessionSsr } from '../modules/auth/session'
 import { redirectToLogin } from '../utils/redirect-to-login'
 
 export const getServerSideProps = withSessionSsr(
