+++ src/pages/login.tsx
@@ -5,7 +5,7 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { LoginForm } from '@/components/forms/login'
 import { Loader } from '@/components/loader/loader'
-import { useLogin } from '@/lib/auth'
+import { useLogin } from '@/modules/auth/hooks'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => {
   const translations = await serverSideTranslations(locale || 'pl', [
