+++ src/pages/players/edit/[slug].tsx
@@ -5,7 +5,7 @@ import { ErrorContent } from '@/components/error/error-content'
 import { EditPlayerForm } from '@/components/forms/player/edit-player'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useCountriesList } from '@/modules/countries/hooks'
 import { usePlayerPositionsList } from '@/modules/player-positions/hooks'
 import { useUpdatePlayer } from '@/modules/players/hooks'
