+++ src/pages/register.tsx
@@ -5,7 +5,7 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { RegisterForm } from '@/components/forms/register'
 import { Loader } from '@/components/loader/loader'
-import { useRegister } from '@/lib/auth'
+import { useRegister } from '@/modules/auth/hooks'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => {
   const translations = await serverSideTranslations(locale || 'pl', [
