+++ src/pages/teams/[slug].tsx
@@ -7,12 +7,12 @@ import { ErrorContent } from '@/components/error/error-content'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { CompetitionParticipationsTable } from '@/components/tables/competition-participations'
 import { CompetitionParticipationsTableRow } from '@/components/tables/rows/competition-participations-row'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useCompetitionParticipations } from '@/modules/competition-participations/hooks'
+import { CompetitionParticipationsSortBy } from '@/modules/competition-participations/types'
 import { TeamDto } from '@/modules/teams/types'
 import { getTeamBySlug } from '@/services/api/methods/teams'
 import { ApiError } from '@/types/common'
-import { CompetitionParticipationsSortBy } from '@/modules/competition-participations/types'
 import { useTable } from '@/utils/hooks/use-table'
 import { redirectToLogin } from '@/utils/redirect-to-login'
