+++ src/pages/teams/edit/[slug].tsx
@@ -5,7 +5,7 @@ import { ErrorContent } from '@/components/error/error-content'
 import { EditTeamForm } from '@/components/forms/team/edit-team'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useClubsList } from '@/modules/clubs/hooks'
 import { useUpdateTeam } from '@/modules/teams/hooks'
 import { TeamDto } from '@/modules/teams/types'
