+++ src/pages/teams/index.tsx
@@ -10,7 +10,7 @@ import { ConfirmationModal } from '@/components/modals/confirmation-modal'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { TeamsTableRow } from '@/components/tables/rows/teams-row'
 import { TeamsTable } from '@/components/tables/teams'
-import { withSessionSsr } from '@/lib/session'
+import { withSessionSsr } from '@/modules/auth/session'
 import { useClubsList } from '@/modules/clubs/hooks'
 import { useCompetitionGroupsList } from '@/modules/competition-groups/hooks'
 import { useCompetitionsList } from '@/modules/competitions/hooks'
